requirements.txt
Flask==0.12
gunicorn==19.7.0
python-dotenv==0.5.1
